requirements.txt
flask
python-dotenv
paho-mqtt